% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderizeAPI.R
\name{checkLanguageCountryCodes}
\alias{checkLanguageCountryCodes}
\title{Check country code.
Ensures that any code specified is recognized by genderize.io.}
\usage{
checkLanguageCountryCodes(countryCode)
}
\description{
Check country code.
Ensures that any code specified is recognized by genderize.io.
}
\keyword{internal}
