% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_author_info.R
\name{get_author_info}
\alias{get_author_info}
\title{get_author_info}
\usage{
get_author_info(df)
}
\arguments{
\item{df}{data frame containing author, title and date}
}
\value{
dataframe with the first and last name of the cited authors
}
\description{
get_author_info
}
\examples{
file_path <- system.file("test-data", "test_citations_table2.csv", package = "CitationProfileR")
sample_data_frame <- read.csv(file_path)
get_author_info(sample_data_frame)
}
