% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location.R
\name{get_location}
\alias{get_location}
\title{get_location}
\usage{
get_location(citation_data, affiliation_col_name = "affiliation.name")
}
\arguments{
\item{citation_data}{Data frame containing affiliations. If using Shiny App, the data will come from \code{crossref_api()} function}

\item{affiliation_col_name}{String name of the affiliations column. The default col name is "affiliation.name".}
}
\value{
Dataframe with country names and country codes based on the inputted affiliation
}
\description{
get_location
}
\examples{
file_path <- system.file("test-data", "test_data_crossref_results.csv", package = "CitationProfileR")
sample_data_frame <- read.csv(file_path)
get_location(sample_data_frame)
\dontrun{
get_location(sample_data_frame, "affiliations_column")
}
}
