% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderizeAPI.R
\name{guess_gender}
\alias{guess_gender}
\title{Guess names' genders}
\usage{
guess_gender(name, key = "ucThdyFEbbTRNp2EhSx4UUn3kMKMThqthcnZ", cache = FALSE)
}
\arguments{
\item{name}{A name to look up.}

\item{apiKey}{An optional API key for https://gender-api.com/.}
}
\description{
This function uses the https://gender-api.com/ API to supply estimates of the gender for one name.
}
\examples{
guess_gender("Natalie", key = "ucThdyFEbbTRNp2EhSx4UUn3kMKMThqthcnZ")
}
\keyword{internal}
