% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderizeAPI.R
\name{lookupNameVectorGenderize}
\alias{lookupNameVectorGenderize}
\title{Look up a vector of names on https://gender-api.com/.}
\usage{
lookupNameVectorGenderize(
  nameVector,
  key = "ucThdyFEbbTRNp2EhSx4UUn3kMKMThqthcnZ"
)
}
\description{
Look up a vector of names on https://gender-api.com/.
}
\keyword{internal}
