% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_pdf_refs.R
\name{parse_pdf_refs}
\alias{parse_pdf_refs}
\title{Parse PDF references}
\usage{
parse_pdf_refs(
  file,
  save_to_file = NULL,
  GROBID_URL = "https://kermitt2-grobid.hf.space"
)
}
\arguments{
\item{file}{path to the PDF file}

\item{save_to_file}{path to the output file}

\item{GROBID_URL}{GROBID URL}
}
\value{
dataframe with references
}
\description{
Parse references in a PDF file using GROBID
}
\examples{
file_path <- system.file("test-data", "Wallrich_et_al_2020.pdf", package = "CitationProfileR")
parse_pdf_refs(file_path)
\dontrun{
parse_pdf_refs(file_path, "file.bib")
parse_pdf_refs(file_path, "file.bib", "https://kermitt2-grobid.hf.space")
}
}
