% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_gender.R
\name{replace_key}
\alias{replace_key}
\title{Users can replace the default API key with their own for the gender-api from https://gender-api.com/}
\usage{
replace_key(key)
}
\arguments{
\item{key}{A user's API key for https://gender-api.com/.}
}
\description{
Users can replace the default API key with their own for the gender-api from https://gender-api.com/
}
\examples{
\dontrun{
#replace_key(key)
}
}
\keyword{internal}
