% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\docType{data}
\name{sample_data_frame}
\alias{sample_data_frame}
\title{Sample citation data with affiliations}
\format{
\subsection{\code{sample_data_frame}}{

A data frame with 104 rows and 17 columns:
\describe{
\item{X}{Row number}
\item{OG_Author}{The original author full name}
\item{OG_Title}{The original title of the paper}
\item{Date}{The date when the paper was published}
\item{Year}{Year the paper was published}
\item{OG_doi}{DOI of the paper}
\item{index}{Index of the entry}
\item{title}{The title of the paper}
\item{given}{The author's given first name}
\item{family}{The author's last name}
\item{sequence}{The sequence of the author's contribution}
\item{affiliation.name}{The name of the affiliations the authors are affiliated with}
\item{affiliation1.name}{The second affiliation}
\item{affiliation2.name}{The third affiliation}
\item{ORCID}{The ORCID ID for the author}
\item{name}{name}
...
}
}
}
\source{
https://jspp.psychopen.eu/index.php/jspp/article/view/5231
}
\usage{
sample_data_frame
}
\description{
Sample citation data with affiliations
}
\keyword{datasets}
