% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_name_check.R
\name{first_name_check}
\alias{first_name_check}
\title{First Name Check}
\usage{
first_name_check(df)
}
\arguments{
\item{df}{A data frame that has the extracted citations from the pdf using GROBID}
}
\value{
a list with two data frames: The first index is entries that have abreviated authors and the second has author's first name
}
\description{
First Name Check
}
