% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_gender.R
\name{guess_gender}
\alias{guess_gender}
\title{Guess a names' gender}
\usage{
guess_gender(name, countrycode = NA, key = NA, cache = FALSE)
}
\arguments{
\item{name}{A name to look up.}

\item{key}{An optional API key for https://gender-api.com/.}

\item{cache}{(logical) local cache to avoid similar request combinations repeating}
}
\description{
This function uses the https://gender-api.com/ API to supply estimates of the gender for one name.
}
\examples{
guess_gender("Rithika", countrycode = "DE", key = "9KyNwgkTTtV4tqgzUWCMXoC6lSNz2ASzTtTd")
}
\keyword{internal}
