% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderizeAPI.R
\name{guessGender}
\alias{guessGender}
\title{Guess names' genders}
\usage{
guessGender(name, apiKey = NA)
}
\arguments{
\item{name}{A name to look up.}

\item{apiKey}{An optional API key for https://gender-api.com/.}
}
\description{
This function uses the https://gender-api.com/ API to supply estimates of the gender for one name.
}
\examples{
guessGender("Natalie")
}
