% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderizeAPI.R
\name{lookupNameVectorGenderize}
\alias{lookupNameVectorGenderize}
\title{Look up a vector of names on genderize.io.}
\usage{
lookupNameVectorGenderize(
  nameVector,
  countryCode = NA,
  apiKey = "7dffb3b373a9efd349a01c214e59e00b"
)
}
\arguments{
\item{nameVector}{A vector containing one or more names to look up.}

\item{countryCode}{An optional ISO 3166-1 alpha-2 country code.}

\item{apiKey}{An optional API key for genderize.io.}
}
\description{
This function actually implements the genderize.io API. Can only query 10
names at a time.
}
\keyword{internal}
